+++ ts/core/data/index.ts
@@ -1,3 +1,3 @@
 export * from './proxy';
 export * from './model';
-// export * from './store';
+export * from './store';